Move code for reading process PACs into input/process/pac.rs
